Rename integration test classes
