+++ cab-app-rides-service/src/integration-test/java/by/arvisit/cabapp/ridesservice/util/RideIntegrationTestData.java
